userMicroService - 기능 추가 : findAll, find(id) GateWay연동

+++ userservice/src/main/java/com/example/userservice/repository/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface UserRepository  {
     void save(UserEntity userEntity);
-    UserEntity find(int userId);
+    UserEntity find(Long userId);
     List<UserEntity> findAll();
 
 }
